% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeu_grad.R
\name{coeu_grad}
\alias{coeu_grad}
\title{Auxiliary function for gradients  and delta2}
\usage{
coeu_grad(mEU2, mEU1, mMS2, mMS1, time2, time1)
}
\arguments{
\item{mEU2}{average at time 2, EU}

\item{mEU1}{average at time 1, EU}

\item{mMS2}{average at time 2, Member State}

\item{mMS1}{average at time 1, Member State}

\item{time2}{time 2}

\item{time1}{time 1}
}
\value{
a list with components time length, grad of member state,
         grad of EU average and the delta squared difference at a pair
         of times.
}
\description{
Intermediate calculation to define patterns.
}
