% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeu_grad.R
\name{coeu_gradV}
\alias{coeu_gradV}
\title{Auxiliary function to provide a different object as input}
\usage{
coeu_gradV(mEU, mMS, time)
}
\arguments{
\item{mEU}{averages at time1 and time 2}

\item{mMS}{indicator for a member country at time1 and time2}

\item{time}{the two times considered, sorted in ascending order}
}
\value{
a list with components time length, grad of member state,
         grad of EU average and the delta squared difference at a pair
         of times.
}
\description{
See  function coeu_grad for details.
}
