% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compo_cond_EUS.R
\name{compo_cond_EUS}
\alias{compo_cond_EUS}
\title{Auxiliary function compo_cond_EUS}
\usage{
compo_cond_EUS(myScarica)
}
\arguments{
\item{myScarica}{a bulk downloaded tibble from Eurostat}
}
\value{
a tag based on indicator-specific  conditioning variables
}
\description{
Not exported
}
