% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergEU-package.R
\docType{package}
\name{convergEU-package}
\alias{convergEU}
\alias{convergEU-package}
\title{convergEU: Monitoring Convergence of EU Countries}
\description{
Indicators and measures by country and time describe what happens at economic and social levels. This package provides functions to calculate several measures of convergence after imputing missing values. The automated downloading of Eurostat data, followed by the production of country fiches and indicator fiches, makes possible to produce automated reports. The Eurofound report (\doi{10.2806/68012}) "Upward convergence in the EU: Concepts, measurements and indicators", 2018, is a detailed presentation of convergence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.eurofound.europa.eu/system/files/2022-04/introduction-to-the-convergeu-package-0.6.4-tutorial-v2-apr2022.pdf}
  \item \url{https://www.eurofound.europa.eu/en/publications/eurofound-paper/2020/monitoring-upward-convergence-eu-r-convergeu-package}
  \item \url{https://www.eurofound.europa.eu/en/publications/2018/upward-convergence-eu-concepts-measurements-and-indicators}
  \item \url{https://www.ajs.or.at/index.php/ajs/article/view/1468}
}

}
\author{
\strong{Maintainer}: Eleonora Peruffo \email{eleonora.peruffo@eurofound.europa.eu}

Authors:
\itemize{
  \item Federico M. Stefanini \email{federico.stefanini@unimi.it} [architect]
}

Other contributors:
\itemize{
  \item Massimiliano Mascherini \email{massimiliano.mascherini@eurofound.europa.eu} [architect]
  \item Nedka Nikiforova [contributor]
  \item Chiara Litardi [contributor]
  \item Berta Mizsei [contributor]
  \item Ricardo Simon-Carbajo [contributor]
  \item Romila Ghosh [contributor]
  \item Andres Suarez-Cetrulo [contributor]
}

}
\keyword{internal}
