% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffQQmu.R
\name{diffQQmu}
\alias{diffQQmu}
\title{Auxiliary function diffQQmu}
\usage{
diffQQmu(vettore)
}
\arguments{
\item{vettore}{a vector of values}
}
\value{
Interquartile range divided by the mean
}
\description{
Not exported
}
