% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dow_soc_scor_boa.R
\name{dow_soc_scor_boa}
\alias{dow_soc_scor_boa}
\title{Downloader of social scoreboard indicators}
\usage{
dow_soc_scor_boa(fromTime = 1999, toTime = 2018, rm.EU = FALSE)
}
\arguments{
\item{fromTime}{starting time}

\item{toTime}{ending time}

\item{rm.EU}{is TRUE remove all variables (columns) starting with
"EU" and "EA", that is averages for different groups of countries.}
}
\value{
a list with as many components as indicators.
}
\description{
This is an "envelope function" to automate the download from Eurostat
of all the indicators involved in the social scoreboard.
}
\details{
Note that the downloaded datasets may have auxiliary columns to be later removed
and they may contain missing values, thus before further calculation taking place,
imputation or truncation of missing values must be performed.
Extra columns include EU12 and other similar weighted averages.
}
\references{
{ \url{https://local.disia.unifi.it/stefanini/RESEARCH/coneu/tutorial-conv.html}}
}
