% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_20_64_MS.R
\docType{data}
\name{emp_20_64_MS}
\alias{emp_20_64_MS}
\title{Dataset emp_20_64_MS}
\format{
A data frame with 17 rows and 29 variables
}
\usage{
data(emp_20_64_MS)
}
\description{
Source data provided by Eurofound, and
reshaped so that the first column is time
and all the other 28 columns are employment values of Member States. The first column refers to the time variable
(e.g. years); the remaining 28 columns refer to the Member States; each Member State is identified through its
corresponding country code accessible by invoking *convergEU_glb()$Eurozone$memberStates*.
}
\examples{
data(emp_20_64_MS)
head(emp_20_64_MS)
names(emp_20_64_MS)

}
\keyword{datasets}
