% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_2_patt_59.R
\name{map_2_patt_59}
\alias{map_2_patt_59}
\title{Values to patterns}
\usage{
map_2_patt_59(vaMS, vaEU, vaT, remap = FALSE)
}
\arguments{
\item{vaMS}{member state values  sorted in ascending order by time.}

\item{vaEU}{EU values sorted in ascending order by time.}

\item{vaT}{sorted pair of times.}

\item{remap}{is FALSE for the original numerical labelling of patterns
otherwise TRUE to map to old numerical correspondence.}
}
\value{
a number referring to  pattern whose label depends on the
         indicator type as originally produced in the technical report.
}
\description{
Gradients values and Delta2 are mapped to one pattern (string and number).
 See Eurofound 2018 report.
 
 
 In the mapping table within this function
 a pattern number 1-59 is assigned for each combination of EU 
 average vs MS average and their differences called delta_endpoints.
 Code NA is left to indicate not relevant features.
 In version 0.6.4The patterns have been updated from the original 12 to 59 
 to cover all possibile combinations.
 Code 60 stands for "to be visually inspected". See vignette for further info.
}
