% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_pattern_39.R
\name{ms_pattern_39}
\alias{ms_pattern_39}
\title{Find patterns for all countries}
\usage{
ms_pattern_39(myTB, timeName = "time")
}
\arguments{
\item{myTB}{a dataset (tibble) for an indicator, time by countries.
The first and last time are respectively the first
and last rows of the dataset,
which must be  time sorted.}

\item{timeName}{a string with name of the time variable}
}
\value{
the type of pattern
}
\description{
The input is a time by countries dataset where all countries contributing
to the average must be present. The expanded set of qualitative equivalence
classes is made by 39 patterns.
}
\details{
This is an expanded  implementation recently submitted for publications.
}
