% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhete1999.R
\name{myhete1999}
\alias{myhete1999}
\title{Auxiliary function myhete1999}
\usage{
myhete1999(vettore)
}
\arguments{
\item{vettore}{a vector of values}
}
\value{
coefficient of variation
}
\description{
Not exported
}
