% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Auxiliary function for membership}
\usage{
not_in(values, set_collection)
}
\arguments{
\item{values}{one or more values}

\item{set_collection}{a collection of values}
}
\value{
TRUE if not within or FALSE otherwise
}
\description{
A fast check if one or more values are outside a set.
}
\examples{

val<-c(1,2,3,5)
mycol<-c(7,8)
not_in(val,mycol)

val1<-c(1,2,3,5)
mycol1<-c(3,5)
not_in(val1,mycol1)

val2<-c("FR", "IT", "LU")
mycol2<-c("FR", "ES")
not_in(val2,mycol2)

}
\references{
{\url{https://local.disia.unifi.it/stefanini/RESEARCH/coneu/tutorial-conv.html}}
}
