% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_legend.R
\name{patt_legend}
\alias{patt_legend}
\title{Graphical legend about time patterns}
\usage{
patt_legend(indiType = "highBest")
}
\arguments{
\item{indiType}{a string equal to "highBest" or "lowBest"
to select a type of indicator.}
}
\value{
a ggplot object to be plotted using grid.arrange()
        function.
}
\description{
A 5 by 4 plot showing patterns of change along time is made
and returned a ggplot object.
}
\examples{

require(gridExtra)
refGGpat2 <- patt_legend(indiType="lowBest")

refGGpat3 <- patt_legend(indiType="highBest")



}
\references{
{\url{https://local.disia.unifi.it/stefanini/RESEARCH/coneu/tutorial-conv.html}}
}
