% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_legend_39.R
\name{patt_legend_39}
\alias{patt_legend_39}
\title{Graphical legend about time patterns}
\usage{
patt_legend_39()
}
\value{
a list of 2 ggplot objects 20 plus 19 patterns.
}
\description{
A plot showing patterns of change along time is made
and returned as a list of ggplot objects.
}
\examples{
require(ggplot2)
refGGpat <- patt_legend_39()

 

}
\references{
{\url{https://local.disia.unifi.it/stefanini/RESEARCH/coneu/tutorial-conv.html}}
}
